+++ dirty_cat/tests/test_minhash_encoder.py
@@ -1,46 +1,39 @@
 import random
-import time
 from string import ascii_lowercase
 
 import numpy as np
 import pandas as pd
 import pytest
-from sklearn.datasets import fetch_20newsgroups
 
 from dirty_cat import MinHashEncoder
-from dirty_cat.datasets import fetch_employee_salaries
 
 
-def test_MinHashEncoder(n_sample: int = 70) -> None:
-    X_txt = fetch_20newsgroups(subset="train")["data"]
-    X = np.array(X_txt[:n_sample])[:, None]
+@pytest.mark.parametrize(
+    "hashing, minmax_hash", [("fast", True), ("fast", False), ("murmur", False)]
+)
+def test_MinHashEncoder(hashing, minmax_hash) -> None:
+    X = np.array(["al ice", "b ob", "bob and alice", "alice and bob"])[:, None]
+    # Test output shape
+    encoder = MinHashEncoder(n_components=2, hashing=hashing)
+    encoder.fit(X)
+    y = encoder.transform(X)
+    assert y.shape == (4, 2), str(y.shape)
+    assert len(set(y[0])) == 2
 
-    for minmax_hash in [True, False]:
-        for hashing in ["fast", "murmur"]:
-            if minmax_hash and hashing == "murmur":
-                pass  # not implemented
+    # Test that using the same seed returns the same output
+    encoder2 = MinHashEncoder(2, hashing=hashing)
+    encoder2.fit(X)
+    y2 = encoder2.transform(X)
+    np.testing.assert_array_equal(y, y2)
 
-            # Test output shape
-            encoder = MinHashEncoder(n_components=50, hashing=hashing)
-            encoder.fit(X)
-            y = encoder.transform(X)
-            assert y.shape == (n_sample, 50), str(y.shape)
-            assert len(set(y[0])) == 50
-
-            # Test same seed return the same output
-            encoder = MinHashEncoder(50, hashing=hashing)
-            encoder.fit(X)
-            y2 = encoder.transform(X)
-            np.testing.assert_array_equal(y, y2)
-
-            # Test min property
-            if not minmax_hash:
-                X_substring = [x[: x.find(" ")] for x in X[:, 0]]
-                X_substring = np.array(X_substring)[:, None]
-                encoder = MinHashEncoder(50, hashing=hashing)
-                encoder.fit(X_substring)
-                y_substring = encoder.transform(X_substring)
-                np.testing.assert_array_less(y - y_substring, 0.0001)
+    # Test min property
+    if not minmax_hash:
+        X_substring = [x[: x.find(" ")] for x in X[:, 0]]
+        X_substring = np.array(X_substring)[:, None]
+        encoder3 = MinHashEncoder(2, hashing=hashing)
+        encoder3.fit(X_substring)
+        y_substring = encoder3.transform(X_substring)
+        np.testing.assert_array_less(y - y_substring, 0.001)
 
 
 def test_multiple_columns() -> None:
@@ -79,28 +72,42 @@ def test_input_type() -> None:
     enc.fit_transform(X)
 
 
-def profile_encoder(hashing: str = "fast", minmax_hash: bool = False) -> float:
-    # not a unit test
-    employee_salaries = fetch_employee_salaries()
-    df = employee_salaries.X
-    X = df[["employee_position_title"]]
-    t0 = time.time()
-    enc = MinHashEncoder(n_components=50, hashing=hashing, minmax_hash=minmax_hash)
+@pytest.mark.parametrize(
+    "hashing, minmax_hash", [("fast", True), ("fast", False), ("murmur", False)]
+)
+def test_encoder_params(hashing, minmax_hash) -> None:
+    MAX_LIMIT = 255  # extended ASCII Character set
+    i = 0
+    str_list = []
+    for i in range(100):
+        random_string = "aa"
+        for _ in range(100):
+            random_integer = random.randint(0, MAX_LIMIT)
+            random_string += chr(random_integer)
+            if random_integer < 50:
+                random_string += "  "
+        i += 1
+        str_list += [random_string]
+    X = np.array(str_list).reshape(100, 1)
+    enc = MinHashEncoder(
+        n_components=50, hashing=hashing, minmax_hash=minmax_hash, ngram_range=(3, 3)
+    )
     enc.fit(X)
     y = enc.transform(X)
     assert y.shape == (len(X), 50)
-    eta = time.time() - t0
-    return eta
+    X2 = np.array([["a", "", "c"]]).T
+    y2 = enc.transform(X2)
+    assert y2.shape == (len(X2), 50)
 
 
-@pytest.mark.parametrize(
-    "input_type, missing, hashing",
-    [
-        ["numpy", "error", "fast"],
-        ["pandas", "zero_impute", "murmur"],
-        ["numpy", "zero_impute", "fast"],
-    ],
-)
+input_types = ["numpy", "pandas"]
+missings = ["error", "zero_impute", "aaa"]
+hashings = ["fast", "murmur", "aaa"]
+
+
+@pytest.mark.parametrize("input_type", input_types)
+@pytest.mark.parametrize("missing", missings)
+@pytest.mark.parametrize("hashing", hashings)
 def test_missing_values(input_type: str, missing: str, hashing: str) -> None:
     X = ["Red", np.nan, "green", "blue", "green", "green", "blue", float("nan")]
     n = 3
@@ -114,23 +121,38 @@ def test_missing_values(input_type: str, missing: str, hashing: str) -> None:
     encoder = MinHashEncoder(
         n_components=n, hashing=hashing, minmax_hash=False, handle_missing=missing
     )
-    if missing == "error":
-        encoder.fit(X)
-        if input_type in ["numpy", "pandas"]:
-            with pytest.raises(ValueError, match=r"missing values in input"):
-                encoder.transform(X)
-    elif missing == "zero_impute":
-        encoder.fit(X)
-        y = encoder.transform(X)
-        if input_type == "list":
-            assert np.allclose(y[1], y[-1])
-        else:
+
+    if hashing == "aaa":
+        with pytest.raises(ValueError, match=r"Got hashing="):
+            encoder.fit_transform(X)
+    else:
+        if missing == "error":
+            if input_type in ["numpy", "pandas"]:
+                with pytest.raises(
+                    ValueError, match=r"Found missing values in input data; set"
+                ):
+                    encoder.fit_transform(X)
+        elif missing == "zero_impute":
+            y = encoder.fit_transform(X)
             assert np.array_equal(y[1], z)
             assert np.array_equal(y[-1], z)
-    else:
-        with pytest.raises(ValueError, match=r"expected any of"):
-            encoder.fit_transform(X)
+        else:
+            with pytest.raises(ValueError, match=r"Got handle_missing="):
+                encoder.fit_transform(X)
     return
+
+
+def test_missing_values_none():
+    # Test that "None" is also understood as a missing value
+    a = np.array([["a", "b", None, "c"]], dtype=object).T
+
+    enc = MinHashEncoder()
+    d = enc.fit_transform(a)
+    np.testing.assert_array_equal(d[2], 0)
+
+    e = np.array([["a", "b", "", "c"]], dtype=object).T
+    f = enc.fit_transform(e)
+    np.testing.assert_array_equal(f[2], 0)
 
 
 def test_cache_overflow() -> None:
@@ -147,17 +169,3 @@ def test_cache_overflow() -> None:
     y = encoder.fit_transform(raw_data)
 
     assert len(y[y == -1.0]) == 0
-
-
-@pytest.mark.parametrize("idx", range(3))
-def test_min_hash_encoder_hashing_fast_minmax_hash(idx):
-    print(f"{profile_encoder(hashing='fast', minmax_hash=True):.4} seconds")
-
-
-@pytest.mark.parametrize("idx", range(3))
-def test_min_hash_encoder_hashing_fast(idx):
-    print(f"{profile_encoder(hashing='fast'):.4} seconds")
-
-
-def test_min_hash_encoder_mumur():
-    print(f"{profile_encoder(hashing='murmur'):.4} seconds")