+++ examples/01_dirty_categories.py
@@ -1,5 +1,4 @@
 """
-==============================================================
 Dirty categories: machine learning with non normalized strings
 ==============================================================
 
@@ -8,13 +7,13 @@ preparation. In particular categories may appear with many morphological
 variants, when they have been manually input or assembled from diverse
 sources.
 
-Here we look at a dataset on wages [#]_ where the column *Employee
-Position Title* contains dirty categories. On such a column, standard
+Here we look at a dataset on wages [#]_ where the column 'Employee
+Position Title' contains dirty categories. On such a column, standard
 categorical encodings leads to very high dimensions and can lose
 information on which categories are similar.
 
 We investigate various encodings of this dirty column for the machine
-learning workflow, predicting the *current annual salary* with gradient
+learning workflow, predicting the 'Current Annual Salary' with gradient
 boosted trees. First we manually assemble a complex encoder for the full
 dataframe, after which we show a much simpler way, albeit with less fine
 control.
@@ -47,7 +46,14 @@ if used on cross-validated subsets of data.
  .. |Gap| replace::
      :class:`~dirty_cat.GapEncoder`
 
- .. |SE| replace:: :class:`~dirty_cat.SimilarityEncoder`
+ .. |MinHash| replace::
+     :class:`~dirty_cat.MinHashEncoder`
+
+ .. |HGBR| replace::
+     :class:`~sklearn.ensemble.HistGradientBoostingRegressor`
+
+ .. |SE| replace::
+     :class:`~dirty_cat.SimilarityEncoder`
 
  .. |permutation importances| replace::
      :func:`~sklearn.inspection.permutation_importance`
@@ -56,7 +62,7 @@ if used on cross-validated subsets of data.
 # %%
 #
 # The data
-# ========
+# --------
 #
 # We first retrieve the dataset:
 from dirty_cat.datasets import fetch_employee_salaries
@@ -69,7 +75,7 @@ X = employee_salaries.X
 X
 
 # %%
-# and y, our target column (the annual salary)
+# and y, our target column (the annual salary):
 y = employee_salaries.y
 y.name
 
@@ -79,7 +85,7 @@ import pandas as pd
 
 X["date_first_hired"] = pd.to_datetime(X["date_first_hired"])
 X["year_first_hired"] = X["date_first_hired"].apply(lambda x: x.year)
-# Get a mask of the rows with missing values in "gender"
+# Get a mask of the rows with missing values in 'gender'
 mask = X.isna()["gender"]
 # And remove them
 X.dropna(subset=["gender"], inplace=True)
@@ -88,7 +94,7 @@ y = y[~mask]
 # %%
 #
 # Assembling a machine-learning pipeline that encodes the data
-# ============================================================
+# ------------------------------------------------------------
 #
 # The learning pipeline
 # ---------------------
@@ -101,15 +107,15 @@ y = y[~mask]
 # ........................
 #
 # An encoder is needed to turn a categorical column into a numerical
-# representation
+# representation:
 from sklearn.preprocessing import OneHotEncoder
 
 one_hot = OneHotEncoder(handle_unknown="ignore", sparse=False)
 
 # %%
 # We assemble these to apply them to the relevant columns.
-# The ColumnTransformer is created by specifying a set of transformers
-# alongside with the column names on which each must be applied
+# The |ColumnTransformer| is created by specifying a set of transformers
+# alongside with the column names on which each must be applied:
 
 from sklearn.compose import make_column_transformer
 
@@ -125,33 +131,29 @@ encoder = make_column_transformer(
 # Pipelining an encoder with a learner
 # ....................................
 #
-# We will use a HistGradientBoostingRegressor, which is a good predictor
-# for data with heterogeneous columns
+# We will use a |HGBR|,
+# which is a good predictor for data with heterogeneous columns
 # (we need to require the experimental feature for scikit-learn versions
-# earlier than 1.0)
-import sklearn
-from sklearn.utils.fixes import parse_version
+# earlier than 1.0):
+from sklearn.experimental import enable_hist_gradient_boosting
 
-if parse_version(sklearn.__version__) < parse_version("1.0"):
-    from sklearn.experimental import enable_hist_gradient_boosting
-# We can now import the HGBR from ensemble
+# We can now import the |HGBR| from ensemble
 from sklearn.ensemble import HistGradientBoostingRegressor
 
 # We then create a pipeline chaining our encoders to a learner
-
 from sklearn.pipeline import make_pipeline
 
 pipeline = make_pipeline(encoder, HistGradientBoostingRegressor())
 
 # %%
-# The pipeline can be readily applied to the dataframe for prediction
+# The pipeline can be readily applied to the dataframe for prediction:
 pipeline.fit(X, y)
 
 # %%
 # Dirty-category encoding
 # -----------------------
 #
-# The one-hot encoder is actually not well suited to the 'Employee
+# The |OneHotEncoder| is actually not well suited to the 'Employee
 # Position Title' column, as this column contains 400 different entries:
 import numpy as np
 
@@ -159,9 +161,14 @@ np.unique(y)
 
 # %%
 # We will now experiment with encoders specially made for handling
-# dirty columns
+# dirty columns:
 
-from dirty_cat import SimilarityEncoder, TargetEncoder, MinHashEncoder, GapEncoder
+from dirty_cat import (
+    SimilarityEncoder,
+    TargetEncoder,
+    MinHashEncoder,
+    GapEncoder,
+)
 
 encoders = {
     "one-hot": one_hot,
@@ -212,14 +219,14 @@ plt.yticks(size=20)
 plt.tight_layout()
 
 # %%
-# The clear trend is that encoders grasping the similarities in the category
-# (similarity, minhash, and gap) perform better than those discarding it.
+# The clear trend is that encoders grasping similarities between categories
+# (|SE|, |MinHash|, and |Gap|) perform better than those discarding it.
 #
-# SimilarityEncoder is the best performer, but it is less scalable on big
-# data than MinHashEncoder and GapEncoder. The most scalable encoder is
-# the MinHashEncoder. GapEncoder, on the other hand, has the benefit that
-# it provides interpretable features
-# (see :ref:`sphx_glr_auto_examples_03_feature_interpretation_gap_encoder.py`)
+# |SE| is the best performer, but it is less scalable on big
+# data than the |MinHash| and |Gap|. The most scalable encoder is
+# the |MinHash|. On the other hand, the |Gap| has the benefit of
+# providing interpretable features
+# (see :ref:`sphx_glr_auto_examples_02_investigating_dirty_categories.py`)
 #
 # |
 #
@@ -228,7 +235,7 @@ plt.tight_layout()
 # .. _example_super_vectorizer:
 #
 # A simpler way: automatic vectorization
-# ======================================
+# --------------------------------------
 #
 # The code to assemble a column transformer is a bit tedious. We will
 # now explore a simpler, automated, way of encoding the data.
@@ -239,8 +246,8 @@ X = employee_salaries.X
 y = employee_salaries.y
 
 # %%
-# We'll drop the "date_first_hired" column as it's redundant with
-# "year_first_hired".
+# We'll drop the 'date_first_hired' column as it's redundant with
+# 'year_first_hired'.
 X = X.drop(["date_first_hired"], axis=1)
 
 # %%
@@ -269,7 +276,7 @@ pipeline = make_pipeline(
 )
 
 # %%
-# Let's perform a cross-validation to see how well this model predicts
+# Let's perform a cross-validation to see how well this model predicts:
 
 from sklearn.model_selection import cross_val_score
 
@@ -324,7 +331,7 @@ pprint(sup_vec.transformers_)
 # %%
 # This is what is being passed to the |ColumnTransformer| under the hood.
 # If you're familiar with how the latter works, it should be very intuitive.
-# We can notice it classified the columns "gender" and "assignment_category"
+# We can notice it classified the columns 'gender' and 'assignment_category'
 # as low cardinality string variables.
 # A |OneHotEncoder| will be applied to these columns.
 #
@@ -344,11 +351,11 @@ feature_names[:8]
 
 # %%
 # As we can see, it gave us interpretable columns.
-# This is because we used |Gap| on the column "division",
+# This is because we used the |Gap| on the column 'division',
 # which was classified as a high cardinality string variable.
 # (default values, see |SV|'s docstring).
 #
-# In total, we have reasonable number of encoded columns.
+# In total, we have a reasonable number of encoded columns:
 len(feature_names)
 
 
@@ -356,15 +363,15 @@ len(feature_names)
 # Feature importances in the statistical model
 # --------------------------------------------
 #
-# In this section, we will train a regressor, and plot the feature importances
+# In this section, we will train a regressor, and plot the feature importances.
 #
 # .. topic:: Note:
 #
-#    To minimize compute time, use the feature importances computed by the
+#    To minimize computation time, we use the feature importances computed by the
 #    |RandomForestRegressor|, but you should prefer |permutation importances|
-#    instead (which are less subject to biases)
+#    instead (which are less subject to biases).
 #
-# First, let's train the |RandomForestRegressor|,
+# First, let's train the |RandomForestRegressor|:
 
 from sklearn.ensemble import RandomForestRegressor
 
@@ -372,7 +379,7 @@ regressor = RandomForestRegressor()
 regressor.fit(X_train_enc, y_train)
 
 # %%
-# Retrieving the feature importances
+# Retrieving the feature importances:
 
 importances = regressor.feature_importances_
 std = np.std([tree.feature_importances_ for tree in regressor.estimators_], axis=0)
@@ -401,131 +408,7 @@ plt.show()
 # having a permanent, full-time job :)
 #
 #
-# |
-#
-
-# %%
-# Working on large datasets: Target Encoding with cross-validation
-# ================================================================
-# Here we discuss how to apply efficiently the :class:`TargetEncoder` if the
-# dataset is large. It may be useful to split the data and encode
-# values on a subset. Overfitting is avoided and we may get
-# better test results. This can be done easily by setting the ``cross_val``
-# parameter of the :class:`TargetEncoder` to True.
-
-# It is also possible to choose the number of outer and inner folds
-# into which the data will be splitted (using the ``n_folds`` and
-# ``n_inner_folds`` parameters).
-
-# %%
-# Data Importing and preprocessing
-# --------------------------------
-#
-# We first get the road safety dataset, containing data from
-# road accidents in Great Britain in 1979, with more than
-# 360 000 observations :
-from dirty_cat.datasets import fetch_road_safety
-
-road_safety = fetch_road_safety()
-# X, our explanatory variables
-X = road_safety.X
-X
-# and y, our target column (the driver's sex)
-y = road_safety.y
-
-# %%
-# Now, let's carry out some basic preprocessing:
-# Keep only the columns that will be used:
-col_to_use = [
-    "Age_of_Driver",
-    "Age_of_Vehicle",
-    "Day_of_Week",
-    "Speed_limit",
-    "Weather_Conditions",
-    "Local_Authority_(Highway)",
-]
-X = X[col_to_use]
-# Drop the lines that contained missing values in X and y
-for col in col_to_use:
-    mask = X.isna()[col]
-    X.dropna(subset=[col], inplace=True)
-    y = y[~mask]
-X.reset_index(drop=True, inplace=True)
-y.reset_index(drop=True, inplace=True)
-# There are more than 208 000 observations left.
-
-# %%
-# Defining the encoders
-# ---------------------
-
-# We create our encoders. In this example we will
-# compare the ``OneHotEncoder`` to the ``MinHasEncoder``
-# and the ``TargetEncoder``, used with or without
-# cross-validated encoding.
-one_hot = OneHotEncoder(handle_unknown="ignore", sparse=False)
-minhash = MinHashEncoder(n_components=100)
-target = TargetEncoder(clf_type="binary-clf", handle_unknown="ignore", cross_val=False)
-target_cv = TargetEncoder(
-    handle_unknown="ignore", cross_val=True, n_folds=4, n_inner_folds=3
-)
-encoders = {
-    "one-hot": one_hot,
-    "minhash": minhash,
-    "target": target,
-    "target-cv": target_cv,
-}
-
-# %%
-# The results
-# ------------
-# We build the pipeline as before and look at the results
-# We will use a HistGradientBoostingClassifier :
-from sklearn.ensemble import HistGradientBoostingClassifier
-from sklearn.model_selection import cross_validate
-
-all_scores = dict()
-for name, method in encoders.items():
-    encoder = make_column_transformer(
-        (one_hot, ["Day_of_Week", "Weather_Conditions", "Speed_limit"]),
-        ("passthrough", ["Age_of_Driver", "Age_of_Vehicle"]),
-        # Last but not least, our dirty column
-        (method, ["Local_Authority_(Highway)"]),
-        remainder="drop",
-    )
-    pipeline = make_pipeline(encoder, HistGradientBoostingClassifier())
-    scores = cross_validate(pipeline, X, y)
-    test_scores = scores["test_score"]
-    print(f"{name} encoding")
-    print(
-        f"r2 score:  mean: {np.mean(test_scores):.3f}; std: {np.std(test_scores):.3f}\n"
-    )
-    all_scores[name] = scores
-# The results show that the :class:`TargetEncoder` performs best
-# if the data are split into folds that will then determine the
-# encoded values. It outperforms also the ``OneHotEncoder``.
-
-# %%
-# Plot a summary figure
-# ---------------------
-fit_times = dict()
-test_results = dict()
-for enc in encoders.keys():
-    fit_times[enc] = all_scores[enc]["fit_time"]
-    test_results[enc] = all_scores[enc]["test_score"]
-_, (ax1, ax2) = plt.subplots(nrows=2, figsize=(4, 3))
-seaborn.boxplot(data=pd.DataFrame(test_results), orient="h", ax=ax1)
-ax1.set_xlabel("Prediction accuracy", size=16)
-[t.set(size=16) for t in ax1.get_yticklabels()]
-seaborn.boxplot(data=pd.DataFrame(fit_times), orient="h", ax=ax2)
-ax2.set_xlabel("Computation time", size=16)
-[t.set(size=16) for t in ax2.get_yticklabels()]
-plt.tight_layout()
-print(test_results)
-
-
-# %%
-#
-# .. topic:: The SuperVectorizer automates preprocessing
+# .. topic:: The |SV| automates preprocessing
 #
 #   As this notebook demonstrates, many preprocessing steps can be
 #   automated by the |SV|, and the resulting pipeline can still be