+++ CHANGES.rst
@@ -1,13 +1,36 @@
 .. currentmodule:: dirty_cat
 
 Release 0.4.0
-=============
+==============
 
-Minor changes
+Major changes
 -------------
+
+* New experimental feature: joining tables using :func:`fuzzy_join` by approximate key matching. Matches are based
+  on string similarities and the nearest neighbors matches are found for each category.
+
+* **datasets.fetching**: contains a new function :func:`fetch_world_bank_indicator` that can be
+    used to download any indicator from the World Bank Open Data platform. It only needs the
+    indicator ID that can be found on the website.
 
 * Unnecessary API has been made private: everything (files, functions, classes)
   starting with an underscore shouldn't be imported in your code.
+
+* :class:`TargetEncoder` can now perform cross-validation loops to encode categories, thus avoiding
+  overfitting. This is done using the `cross_val` parameter, and the number of splits of the
+  dataset are set with the `n_folds` and `n_inner_folds` parameters.
+
+Minor changes
+-------------
+* Removed example `Fitting scalable, non-linear models on data with dirty categories`.
+
+* :class:`MinHashEncoder`'s `minhash` method is no longer public.
+
+Bug fixes
+---------
+
+* :class:`MinHashEncoder` now considers `None` and empty strings as missing values, rather
+  than raising an error.
 
 Release 0.3.0
 =============
@@ -45,10 +68,6 @@ Major changes
     The :class:`SimilarityEncoder` now exclusively uses ``ngram`` for similarities,
     and the `similarity` parameter is deprecated. It will be removed in 0.5.
 
-* :class:`TargetEncoder` can now perform cross-validation loops to encode categories, thus avoiding
-    overfitting. This is done using the `cross_val` parameter, and the number of splits of the
-    dataset are set with the `n_folds` and `n_inner_folds` parameters.
-
 Notes
 -----
 
@@ -62,7 +81,7 @@ Notes
 Release 0.2.2
 =============
 
-Bug-fixes
+Bug fixes
 ---------
 
 * Fixed a bug in the :class:`SuperVectorizer` causing a `FutureWarning`